Updated Package
Updated Laravel to 8.10 an InertiaLaravel to 0.3 and all required configuration
