+++ app/Providers/AppServiceProvider.php
@@ -8,13 +8,10 @@ use League\Glide\Server;
 use Carbon\CarbonImmutable;
 use Illuminate\Support\Collection;
 use Illuminate\Pagination\UrlWindow;
-use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Date;
 use Illuminate\Support\Facades\Request;
-use Illuminate\Support\Facades\Session;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Pagination\LengthAwarePaginator;
-use App\Models\User;
 
 class AppServiceProvider extends ServiceProvider
 {
@@ -25,7 +22,6 @@ class AppServiceProvider extends ServiceProvider
      */
      public function register()
      {
-          $this->registerInertia();
           $this->registerGlide();
           $this->registerLengthAwarePaginator();
      }
@@ -38,61 +34,6 @@ class AppServiceProvider extends ServiceProvider
      public function boot()
      {
           Date::use(CarbonImmutable::class);
-     }
-
-     public function registerInertia(){
-          Inertia::version(function () {
-               return md5_file(public_path('mix-manifest.json'));
-          });
-
-          Inertia::share([
-               'app' => function(){
-                    return [
-                         'name' => config('app.name')
-                    ];
-               },
-               'auth' => function () {
-                    $user = Auth::user() ? User::with(['roles' => function($roles){
-                         return $roles->with('permissions');
-                    }])->find(Auth::user()->id) : null;
-                    $roles = array();
-                    $permissions = array();
-                    if($user){
-                         foreach($user->roles as $r){
-                              array_push($roles, [
-                                   'name' => $r->name,
-                                   'display' => $r->display_name
-                              ]);
-                              foreach($r->permissions as $p){
-                                   array_push($permissions, $p->name);
-                              }
-                         }
-                    }
-
-                    return [
-                         'user' => $user ? [
-                              'id' => $user->id,
-                              'name' => $user->name,
-                              'username' => $user->username,
-                              'email' => $user->email,
-                              'avatar' => ($user->profile_url) ? route('show.image', 'users/'.$user->profile_url) : null, // (Auth::user()->profile_url) ? asset('storage/users/'.Auth::user()->profile_url) : null
-                              'roles' => ($roles) ? $roles : null,
-                              'can' => ($permissions) ? $permissions : null
-                         ] : null,
-                    ];
-               },
-               'flash' => function () {
-                    return [
-                         'success' => Session::get('success'),
-                         'error' => Session::get('error'),
-                    ];
-               },
-               'errors' => function () {
-                    return Session::get('errors')
-                         ? Session::get('errors')->getBag('default')->getMessages()
-                         : (object) [];
-                    },
-          ]);
      }
 
      protected function registerGlide(){
