+++ composer.json
@@ -12,9 +12,9 @@
         "fideloper/proxy": "^4.2",
         "fruitcake/laravel-cors": "^2.0",
         "guzzlehttp/guzzle": "^7.0.1",
-        "inertiajs/inertia-laravel": "^0.2.12",
+        "inertiajs/inertia-laravel": "^0.3",
         "laravel/fortify": "^1.5",
-        "laravel/framework": "^8.0",
+        "laravel/framework": "^8.10",
         "laravel/tinker": "^2.0",
         "santigarcor/laratrust": "^6.2",
         "tightenco/ziggy": "^0.9.4"
