+++ public/19.js
@@ -24,7 +24,6 @@ var Home = function Home() {
   var _usePage = Object(_inertiajs_inertia_react__WEBPACK_IMPORTED_MODULE_2__["usePage"])(),
       auth = _usePage.auth;
 
-  console.log(auth.user);
   return /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement("div", null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(react_helmet__WEBPACK_IMPORTED_MODULE_1__["default"], null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement("title", null, "Dashboard")));
 }; // Persisten layout
 // Docs: https://inertiajs.com/pages#persistent-layouts
